+++ src/util.cc
@@ -1,0 +1,83 @@
+#include "util.h"
+
+void deleteArray(const Napi::Env &env, unsigned char *ptr)
+{
+    delete[] ptr;
+}
+
+std::string copyArrayOrBufferIntoVector(const Napi::Value &val, std::vector<unsigned char> &message)
+{
+    if (val.IsBuffer())
+    {
+        Napi::Buffer<unsigned char> buffer = val.As<Napi::Buffer<unsigned char>>();
+        uint32_t len = buffer.Length();
+        unsigned char *data = buffer.Data();
+        message.assign(data, data + len);
+
+        return "";
+    }
+    else if (val.IsArray())
+    {
+        Napi::Array messageArray = val.As<Napi::Array>();
+        message.reserve(messageArray.Length());
+
+        for (unsigned i = 0; i < messageArray.Length(); i++)
+        {
+            Napi::Value v = messageArray.Get(i);
+            if (!v.IsNumber())
+            {
+                return "unexpected array element in array to send, expecting only integers";
+            }
+            uint32_t b = v.As<Napi::Number>().Uint32Value();
+            message.push_back((unsigned char)b);
+        }
+
+        return "";
+    }
+    else
+    {
+        return "unexpected data to send, expecting an array or buffer";
+    }
+}
+
+WrappedHidHandle::WrappedHidHandle(hid_device *hidHandle) : hid(hidHandle) {}
+WrappedHidHandle::~WrappedHidHandle()
+{
+    if (hid)
+    {
+        hid_close(hid);
+        hid = nullptr;
+    }
+
+    // TODO - discard the jobQueue in a safe manner
+}
+
+void WrappedHidHandle::QueueJob(const Napi::Env &, Napi::AsyncWorker *job)
+{
+    std::unique_lock<std::mutex> lock(jobQueueMutex);
+    if (!isRunning)
+    {
+        isRunning = true;
+        job->Queue();
+    }
+    else
+    {
+        jobQueue.push(job);
+    }
+}
+
+void WrappedHidHandle::JobFinished(const Napi::Env &)
+{
+    std::unique_lock<std::mutex> lock(jobQueueMutex);
+
+    if (jobQueue.size() == 0)
+    {
+        isRunning = false;
+    }
+    else
+    {
+        auto newJob = jobQueue.front();
+        jobQueue.pop();
+        newJob->Queue();
+    }
+}
